Added example item config file
This file is just an example to show the format of the config files. Configs can have sections containing different lists as needed, which can be obtained as Generic Maps.

The actual path of the file will be in the server's plugin folder under the specific plugin's config folder.
e.g. server/plugins/DrillServerMineZPlugin/items.yml